Refactoring of upgrade scripts:
* Upgrade is now divided in 3 separate scripts: download-bundle, upgrade-target and cleanup. This fixes/provides:
  * Windows failing when deleting scripts currently being run.
  * Upgrading a bundle that doesn't have upgrade scripts yet.
  * Upgrading installations without an Internet connection.
  * Possibility of having custom upgrade code in future versions.
