+++ environments.gradle
@@ -116,7 +116,8 @@ task("authoringPostEnvironment") {
             exec {
                 workingDir "${authEnv}/bin"
                 commandLine commandLinePrefix + ["chmod", "+x", "crafter.sh", "startup.sh", "debug.sh", "shutdown.sh",
-                                                 "crafter-setenv.sh", "upgrade/upgrade.sh"]
+                                                 "crafter-setenv.sh", "upgrade/download-bundle.sh",
+                                                 "upgrade/upgrade-target.sh"]
             }
         }
 
@@ -126,9 +127,11 @@ task("authoringPostEnvironment") {
         updateTomcatContext("${authEnv}/bin/apache-tomcat/conf/context.xml")
         updateCatalinaLogging("${authEnv}/bin/apache-tomcat/conf/logging.properties")
         addSolrIndexHome("${authEnv}/bin/solr/server/solr/solr.xml")
-        downloadGrapes("${authEnv}/bin", "upgrade/setup.groovy")
-        downloadGrapes("${authEnv}/bin", "upgrade/upgrade.groovy")
-        downloadGrapes("${authEnv}/bin", "upgrade/cleanup.groovy")
+        // downloadGrapes("${authEnv}/bin", "upgrade/setup.groovy")
+        // downloadGrapes("${authEnv}/bin", "upgrade/upgrade.groovy")
+        // downloadGrapes("${authEnv}/bin", "upgrade/cleanup.groovy")
+        downloadGrapes("${authEnv}/bin", "upgrade/download-bundle.groovy")
+        downloadGrapes("${authEnv}/bin", "upgrade/upgrade-target.groovy")
     }
 }
 
@@ -211,7 +214,7 @@ task("deliveryPostEnvironment") {
                 workingDir "${deliveryEnv}/bin"
                 commandLine commandLinePrefix + ["chmod", "+x", "crafter.sh", "startup.sh", "debug.sh", "shutdown.sh",
                                                  "crafter-setenv.sh", "init-site.sh", "remove-site.sh",
-                                                 "upgrade/upgrade.sh"]
+                                                 "upgrade/download-bundle.sh", "upgrade/upgrade-target.sh"]
             }
         }
 
@@ -223,9 +226,11 @@ task("deliveryPostEnvironment") {
         addSolrIndexHome("${deliveryEnv}/bin/solr/server/solr/solr.xml")
         downloadGrapes("${deliveryEnv}/bin", "init-site.groovy")
         downloadGrapes("${deliveryEnv}/bin", "remove-site.groovy")
-        downloadGrapes("${deliveryEnv}/bin", "upgrade/setup.groovy")
-        downloadGrapes("${deliveryEnv}/bin", "upgrade/upgrade.groovy")
-        downloadGrapes("${deliveryEnv}/bin", "upgrade/cleanup.groovy")
+        // downloadGrapes("${deliveryEnv}/bin", "upgrade/setup.groovy")
+        // downloadGrapes("${deliveryEnv}/bin", "upgrade/upgrade.groovy")
+        // downloadGrapes("${deliveryEnv}/bin", "upgrade/cleanup.groovy")
+        downloadGrapes("${deliveryEnv}/bin", "upgrade/download-bundle.groovy")
+        downloadGrapes("${deliveryEnv}/bin", "upgrade/upgrade-target.groovy")
     }
 }
