+++ resources/crafter/upgrade/cleanup.groovy
@@ -5,4 +5,8 @@ import static utils.ScriptUtils.*
 
 checkDownloadGrapesOnlyMode(getClass())
 
+println "============================================================"
+println "Deleting temp upgrade folder"
+println "============================================================"
+
 NioUtils.deleteDirectory(getUpgradeTmpFolder())
