+++ resources/crafter/utils/ScriptUtils.groovy
@@ -1,6 +1,11 @@
 package utils
 
-@Grab(group='commons-io', module='commons-io', version='2.6')
+@Grapes([
+		@Grab(group='org.apache.commons', module='commons-lang3', version='3.7'),
+		@Grab(group='commons-io', module='commons-io', version='2.6')
+])
+
+import org.apache.commons.lang3.SystemUtils
 import org.apache.commons.io.FilenameUtils
 
 import static utils.EnvironmentUtils.*
@@ -22,6 +27,37 @@ class ScriptUtils {
 			println "Downloading grapes for ${getScriptName(scriptClass)}..."
 
 			System.exit(0)
+		}
+	}
+
+	/**
+	 * Executes a command line process.
+	 */
+	static def executeCommand(workingDir, command, setupCallback = null) {
+		if (SystemUtils.IS_OS_WINDOWS) {
+			command  = ["cmd", "/c"] + command
+		}
+
+		def processBuilder = new ProcessBuilder(command)
+
+		if (workingDir) {
+			processBuilder.directory(workingDir.toFile())
+		}
+
+		if (setupCallback) {
+			setupCallback(processBuilder)
+		}
+
+		processBuilder.redirectOutput(ProcessBuilder.Redirect.INHERIT)
+		processBuilder.redirectError(ProcessBuilder.Redirect.INHERIT)
+
+		def process = processBuilder.start()
+				process.waitFor()
+
+		def exitValue = process.exitValue()
+
+		if (exitValue != 0) {
+			throw new RuntimeException("Process '${command}' exited with non-successful value ${exitValue}")
 		}
 	}
